fetch image names from database
